Struct detour::graph::graph::DetourGraph

pub struct DetourGraph {
    graph: StableDiGraph<Bbox, Vec<[f64; 3]>>,
    roots: Vec<NodeIndex>,
    outliers: StableDiGraph<Bbox, Vec<[f64; 3]>>,
}

Fields
graph: StableDiGraph<Bbox, Vec<[f64; 3]>>roots: Vec<NodeIndex>outliers: StableDiGraph<Bbox, Vec<[f64; 3]>>
Implementations

impl DetourGraph

pub fn new() -> DetourGraph

pub fn set_graph(
&mut self,
new_graph: StableDiGraph<Bbox, Vec<[f64; 3]>>,
root_nodes: Vec<NodeIndex>
)

pub fn get_mut_graph(&mut self) -> &mut StableDiGraph<Bbox, Vec<[f64; 3]>>

pub fn get_graph(&self) -> &StableDiGraph<Bbox, Vec<[f64; 3]>>

pub fn set_outlier_graph(
&mut self,
outlier_graph: StableDiGraph<Bbox, Vec<[f64; 3]>>
)

pub fn to_csv(&self) -> Result<()>

Writes the graph to the output folder. The coordinates of node and edge weights, i.e., bounding boxes and trajectories respectively, are stored in Web Mercator format.

A Visual representation are stored in graph.dot and outlier_graph.dot for the regular graph and the outlier graph, respectively.

The regular graph is stored in graph.json and the outlier graph is stored in outlier_graph.json

Information about splits and merges of nodes/edges are stored in STATS.

Configuration is written to config for reference purposes.

fn merge_node_pair(&mut self, a: NodeIndex, b: NodeIndex) -> NodeIndex

Merges a pair of nodes and returns the new node index

pub fn merge_nodes(&mut self)

fn shrink_node(&mut self, nx: NodeIndex)

Shrink bbox based on points of connected trajectories. (not just endpoints)

fn should_split(&self, nx: NodeIndex) -> bool

fn split_node_at(&mut self, split_node: NodeIndex, splits: &[f64])

fn split_node(&mut self, split_node: NodeIndex)

fn get_temporal_splits(&self, split_node: NodeIndex) -> Vec<f64>

fn get_trjs_in_timeframe(&self, node: NodeIndex) -> Vec<EdgeIndex>

Return edge indices connected to node that potentially violates temporal monotonicity i.e. their trajectories’ temporal start and end lie within bbox of the node.

fn split_bbox(bbox: Bbox, splits: &[f64]) -> Vec<Bbox>

fn reassign_edges(
&mut self,
split_node: NodeIndex,
nodes: &[NodeIndex],
direction: EdgeDirection
)

fn get_root_case(&mut self, nx_a: NodeIndex, nx_b: NodeIndex) -> RootCase

pub fn remove_node(&mut self, nx: NodeIndex)

Removes node from the graph and the root list.

fn get_root_index(&self, nx: NodeIndex) -> Option<usize>

Returns Some(index) of nx in the roots list or None

pub fn node_indices(&self) -> NodeIndices<'_, Bbox>

Allows iteration over nodes.

pub fn edge_weights(&self) -> Vec<&Vec<[f64; 3]>>

Allows iteration over edge weights.

pub fn get_node_weight(&self, nx: NodeIndex) -> Bbox

pub fn set_node_weight(&mut self, nx: NodeIndex, bbox: Bbox)

pub fn edges_directed(
&self,
nx: NodeIndex,
dir: EdgeDirection
) -> Vec<EdgeIndex>

pub fn edge_weight_mut(&mut self, ex: EdgeIndex) -> &mut Vec<[f64; 3]>

fn get_dominant_roots(&self, nx: NodeIndex) -> Vec<usize>

Returns indices of nodes in self.roots from which nx can be reached

fn copy_edges(&mut self, copyer: NodeIndex, copyed: NodeIndex)

fn verify_constraints(&self) -> bool

fn verify_node(&self, nx: NodeIndex) -> (bool, bool, bool)

Checks and returns a tuple with the following for the node, nx: (is_root, is_orphan, is_root_reachable)

fn verify_temporal_monotonicity_old(&self) -> bool

If any edge (a,b) breaks temporal monotonicity -> false.

fn verify_temporal_monotonicity(&self) -> bool

If any edge (a,b) breaks temporal monotonicity -> false.

fn fix_temporal_monotonicity(&mut self)

finds and fixes temporal monotonicity by splitting nodes.

fn find_matching_nodes(&self) -> Option<(NodeIndex, NodeIndex)>

fn root_reachable(&self, target: NodeIndex) -> bool

pub fn add_path(&mut self, path: Path)