Struct detour::StaticDetour

pub struct StaticDetour<T: Function>(_);

A type-safe static detour.

It can only be created using StaticDetourController::initialize.

When this object has been dropped, the detour is freed and the controller can be initialized once again.
It dereferences to GenericDetour so it provides the same functions that it (and Detour) provides.

Beyond this it also provides a set_detour method, enabling the detour to be changed whilst hooked.

To see an example view the macro's page.

Methods

impl<T: Function> StaticDetour<T>

fn set_detour<C>(&mut self, closure: C) where
C: Fn<T::Arguments, Output = T::Output> + Send + 'static,

Changes the detour, regardless of whether the target is hooked or not.