Struct detour::utility::bounding_box::Bbox

pub struct Bbox {
    pub x1: f64,
    pub x2: f64,
    pub y1: f64,
    pub y2: f64,
    pub t1: f64,
    pub t2: f64,
}

Fields
x1: f64x2: f64y1: f64y2: f64t1: f64t2: f64
Implementations

impl Bbox

pub fn new(pts: &[[f64; 3]]) -> Bbox

pub fn nearest_corner(&self, point: &[f64; 3]) -> [f64; 3]

Returns the corner spatially nearest to point The time coordinate of the returned point is copied from point

pub fn overlaps(&self, other: &Self) -> bool

Determine if the bounding boxes overlap spatially and temporally

pub fn overlaps_spatially(&self, other: &Self) -> bool

Determine if the bounding boxes overlap spatially

pub fn insert_point(&mut self, point: &[f64; 3])

Inserts point and expands Bbox if neccesary.

pub fn is_before(&self, other: &Self) -> bool

Determine if other ‘Bbox’ is temporally before ‘self’.

Specifically, for Bboxes x and y; if x ends before y starts, return true

Example:

let a = Bbox::new(&[[0., 0., 0.], [2., 2., 2.]]);
let b = Bbox::new(&[[0., 0., 1.], [2., 2., 3.]]);
let c = Bbox::new(&[[0., 0., 2.5], [2., 2., 4.]]);
assert!(a.is_before(&c));
assert!(!b.is_before(&c));

pub fn contains_point(&self, pt: &[f64; 3]) -> bool

pub fn is_in_temporal(&self, pt: &[f64; 3]) -> bool

pub fn is_in_spatial(&self, pt: &[f64; 3]) -> bool

pub fn get_diameter(&self) -> f64

Returns the squared diameter of the bbox.

pub fn temporal_split(&self, t: f64) -> (Self, Self)

Splits Bbox temporally at t.

Note that we assume a granularity of whole ms. Thus, the first Bbox will contain t and the second will contain t+1ms

pub fn expand(&mut self, meters: f64, minutes: f64)

pub fn verify_spatial(&self) -> bool

Verifies if bbox satisfies the spatial constraints given in config

pub fn verify_temporal(&self) -> bool

Verifies if bbox satisfies the temporal constraints given in config

pub fn verify(&self) -> bool

Verifies if bbox satisfies the constraints given in config

pub fn union(&self, other: &Self) -> Self

pub fn is_single_point(&self) -> bool

pub fn can_contain_trj(&self, trj: &[[f64; 3]]) -> bool

pub fn expand_along_trjs(
&self,
trjs: Vec<Vec<[f64; 3]>>,
t1: Option<f64>,
t2: Option<f64>
) -> Bbox

Expands the bbox along the trajectories. TODO: When building paths, we probably want to merge two nodes if they both can contain the entire trajectory.