Crate detour

Constructs Detour Graph

Command line utility for constructing an acyclic graph that represents condensed movement patterns of the GPX files given as input. Streams are cleaned using a temporal filter which ensures temporal monotonicity.

Example

Say you have a bunch of GPX files located in /home/user/gpx you want to use as input. Pipe the output to the program like:

$ cat /home/user/gpx/* | .detour -o /home/user/output

To use a specific configuration file use the -c or --config option followed by desired configuration file. By default output is written to a folder called Output unless otherwise is specified by using the -o or --output option.

Configuration

Various settings can be adjusted by modifying config.cfg located in the root folder. Read more about Config here.

Re-exports

pub use config::Config;

Modules

arguments

config

coord

graph

parser

time_guard

utility

Various utilities

Structs

BufReader

The BufReader<R> struct adds buffering to any reader.

CHFilter

Applies the convex hull filter described by Adhinugraha et al.

Coord

Holds coordinates in UTM projection.

StopDetector

Detect if the object is stopped. If the movements of the object within a time frame of duration_ms is limited to a geofenced location with a diagonal of diagonal_meters the object is considered to be stopped.

Traits

Read

The Read trait allows for reading bytes from a source.

Functions

main