Crate detour

Constructs Detour Graph

Command line utility for constructing an acyclic graph that represents condensed movement patterns of the GPX files given as input. Streams are cleaned using a temporal filter which ensures temporal monotonicity.

Example

Say you have a bunch of GPX files located in /home/user/gpx you want to use as input. Pipe the output to the program and specify an output folder like so:

$ cat /home/user/gpx/* | .detour /home/user/output

Configuration

Various settings can be adjusted by modifying config.cfg located in the root folder. Read more about Config here.

Modules

data_structures

env

Inspection and manipulation of the process’s environment.

fs

Filesystem manipulation operations.

parser

time_guard

utility

Various utilities

Macros

env

Inspects an environment variable at compile time.

Structs

BufReader

The BufReader<R> struct adds buffering to any reader.

CHFilter

Applies the convex hull filter described by Adhinugraha et al.

Config

The fields can be set using the config file.

Coord

Holds coordinates in UTM projection.

Path

A slice of a path (akin to str).

StopDetector

Detect if the object is stopped. If the movements of the object within a time frame of duration_ms is limited to a geofenced location with a diagonal of diagonal_meters the object is considered to be stopped.

Traits

Read

The Read trait allows for reading bytes from a source.

Functions

main